+++ src/main/java/ch/furchert/homelab/auth/AuthServiceApplication.java
@@ -3,9 +3,11 @@ package ch.furchert.homelab.auth;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @ConfigurationPropertiesScan
+@EnableScheduling
 public class AuthServiceApplication {
 
 	public static void main(String[] args) {