+++ src/main/java/ch/furchert/homelab/auth/config/RsaKeyProperties.java
@@ -4,9 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.core.io.Resource;
-import org.springframework.stereotype.Component;
 
-@Component
 @ConfigurationProperties(prefix = "app.jwt")
 @Getter
 @Setter